+++ src/main/java/uk/gov/hmcts/reform/civil/controllers/citizenui/controller/Example.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.civil.controllers.citizenui.controller;
+
+public class Example {
+}
