--- src/main/java/uk/gov/hmcts/reform/civil/model/citizenui/controller/Example.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.civil.model.citizenui.controller;
-
-public class Example {
-}
