--- src/main/java/uk/gov/hmcts/reform/civil/model/citizenui/repository/Example.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.civil.model.citizenui.repository;
-
-public class Example {
-}
