+++ src/main/java/uk/gov/hmcts/reform/civil/service/citizenui/dashboard/Example.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.civil.service.citizenui.dashboard;
+
+public class Example {
+}
